requirements.txt
alembic==1.13.1
APScheduler==3.11.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
blinker==1.9.0
cachelib==0.13.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
cssselect2==0.7.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.0
Flask-APScheduler==1.13.1
Flask-Login==0.6.2
Flask-Migrate==4.0.7
Flask-Session==0.7.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
greenlet==3.1.1
html5lib==1.1
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.4
lxml==5.3.0
Mako==1.3.8
MarkupSafe==2.1.5
msgspec==0.19.0
oscrypto==1.3.0
pillow==11.0.0
pycparser==2.22
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pypdf==5.1.0
python-bidi==0.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.25
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.12.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.3.0
webencodings==0.5.1
Werkzeug==2.3.8
WTForms==3.2.1
xhtml2pdf==0.2.16
